Write listening port to a file for clients to pick up
As we select the tcp port randomly, write a file with the address we are
listening on for clients to know how to contact us.
Note that we are using a file so that even client on WSL (with interop
enabled) are able to read it. This is then a requirement for Ubuntu Pro.