Implemented distro module
This module concerns itself with the management of distros and their
event loop. This event loops processes tasks in a FIFO queue.

In order to process each task, the a keepawake lock is set on the
distro, then we wait for a connection to the WSL instance to exist. Once
it exists, the distro processes this task. With the task finished
--successfully or not--, we release the keepawake lock.

IF the distro cannot be reached, it is marked as such.

Co-authored-by: Didier Roche-Tolomelli <didrocks@ubuntu.com>