Changed marshaling to use tasks instead of managed task
Distro works with managed tasks but InitialTasks works with the tasks
themselves, so we want to store and load them directly as tasks instead
if as this Managed object that has info relevant only to the distro.

managed.go has also been renamed since the main object is no longer
managed tasks. The Managed object has been renamed managedTask and moved
back to task_managent.go.