Add mockito for more complex tests
With mockito comes the need for codegen through build_runner
It generates the mock classes.
CI is updated to check for outdated mocks
or other machine generated Dart code except gRPC related.
gRPC was already covered.

I understand that mocks incentivize stateful APIs.
But GUI has a lot to do with stateful APIs anyway.
And we likely don't want to spin up a gRPC server to test whether a
button is disabled when it should or other small UI behaviors.