Brings the Either type
Similar to C++ std::expected and Rust std::Result
but backwards, because it's designed from Haskell's Either type.

Useful for error handling when errors are not exceptional situations.
Type has a Monadic interface.
There are other implementations in Dart.
They are all part of bigger libs dedicated to functional programming.
This focuses only in the Either type and nothing else FP alike.