Implements and applies an Environment singleton
So we can override the env variables map for testing.

It's customizability is meant only for testing.
If we need something similar in production
then the singleton will be a hindrance
thus we should drop this in favor of a more robust solution
like dotenv: https://pub.dev/packages/dotenv