Bugfix: TestConnected logged after finishing
Using t.Cleanup instead of defer can be considered a bad practice.
Although very similar, defer is called right before the test finishes,
whereas the cleanup is triggered right after. This means that any
goroutine that may assert or log is susceptible to do so once the test
is over. This'll cause a panic, as seen here:

https://github.com/canonical/ubuntu-pro-for-windows/actions/runs/4565779740/jobs/8057379439?pr=73

These changes fix this issue by releasing resources with defer rather
than cleanup.