Implements the AgentStartupMonitor
Contains the logic to detect the agent running
or request its startup.
On success it creates the AgentApiClient
 and passes to the supplied callback.

Translatable states are offered for client widgets.

The actual agent startup request is not implemented here.
Instead it is a function this receives as a parameter.
The intended implementation must request the native plugin.