Sets up WinRT via CMake on P4wMsStore plugin.
It will contain all Store and App Model APIs, which require WinRT.

CMake FetchContent to fetch NuGet.
NuGet installs CppWinRT in the build tree.
CppWinRT generates the WinRT C++ headers.

Special treatment is required for the FullTrustAppContract.
That is not generated by default.
Requires referring to the metadata file in the cppwinrt.exe CLI.
I struggled A LOT to figure that out.

All that above for the Windows impl only.
Linux impl (mock) to be worked out at next.