Native plugin tests with Google Test
Recipe very similar to flutter official plugins.
The key difference is that the BP is to rely on the example app.
This plugin doesn't have one.
The only client of that plugin is supposed to be the Pro app.

We thus rely on the env var INCLUDE_PRO_PLUGIN_TESTS
to conditionally build the tests.
That build process can be expensive.
It downloads and builds gtest.

Not fully in love with it, but it looks promising as the plugin grows.
There are still quite a lot to add related to the Store.

Small tests were added.