wsl-pro-service/systeminfo: Fixed obtaining LocalAppData
We were relying on variable expansion, which can be done when writing
a command on the shell but not when starting a subprocess from Go.

Furthermore, the Windows path is not preappended to the path passed to
systemd services, so we have to find the executable ourselves. Instead
of using powershell.exe, I used cmd.exe because it has a more stable
path.

The first time we need to use cmd.exe, we scan all /proc/mounts until
we find the mount that contains cmd.exe in the expected path. This path
is cached.