Revert commit a1156345d02419b020d461f3cbac57a62cee33af
This was done after misreading the docs of the GRPC module. The
problematic error is:

    grpc: the server has been stopped.

This error message shows up when you try to perform an action and the
server was stopped already. The reason this was triggered is that the
server takes a while to spin up (particularly with many tests running
concurrently), so we were calling "defer s.Quit" at the end of the test
before a.Run started serving.

The solution is to ensure a.Run has ample time to start serving before
starting the tests.