Stop leaking worker goroutines
The Stop function on distro must wait for the cleanup process to finish.
Otherwise there is no way of gracefully terminating a program with
any guarantees that the distro has finished savin to file.

If we wanted the functionality to quit without waiting, we can use

   go distro.Stop(ctx)

so we do not have any loss of functionality.

Otherwise, tests that need to clean up after they are done will often
have filesystem races (attempting to remove the tmpdir while the distro
is writing its dumpfile in it).