Implement config
The config stores the data necessary to provision new distros (for now,
more data later) and creates provisioning tasks on demand. Its data is
stored in the Windows Registry, which is interacted with via an
interface so that tests can easily swap it with a mock.

In the future, the config manager may need to actively monitor the
registry in order to push updated pro tokens to existing distros. This
is out of scope for now. Hence, I took a simpler read-when-you-need
approach.