Local only accounts don't generate UserInfo
Such as those local users in VMs.
None of the KnownUserProperties produces output.
This we need to realize that the string can be empty.

Yet, the user ID must be hashed with SHA256.
I had forgotten that requirement. It's in the spec.
Good that Hex of SHA256 has a fixed size.