Bugfix: Race allowed worker to keep processing tasks after Stop
The choice between quitting and processing a task hinged on a select
statement. This means that if
a) tasks keep getting pushed faster than they are popped, or
b) there is a task backlog

there is always a chance that the worker will select the task channel
rather than the context.Done.

This fix prevents this race. This caused issues sporadically in tests.