Implements Client GetProToken()
Supplies a provided a user JWT to the CS backend.
Receives a single-field JSON object containing the Pro Token,
which may or may not be new. It'll depend on when this was called.

A higher level component might monitor the MS Store events and use this
to notify the backend about changes in the subscription
such as a new purchase or auto renewal.