Change worker interpretation of NeedsRetry
Used to mean: retry immediately
Now it means: retry on next startup

-- How? --

Tasks are picked up from the queue (a channel) that contains pointers to
these tasks, which are found in a slice. When a task is picked up, it is
popped from the queue, but not from the slice.

When the task finishes it returns an error:
- If the error is nil: we remove the task from the slice
- If the error is NeedsRetryError: we keep it in the slice
- Any other type of error: we remove the task from the slice

After this, the slice is stored to disk. Since the task is NOT in the
queue, it will not be picked up.

Upon restarting the agent, all tasks are loaded from disk and queued,
meaning that it can be picked up once again.