jwtBuf has to be a pointer to pointer to char.
See the signature to the syscall.Lazyproc.Call:

func (p *LazyProc) Call(a ...uintptr) (r1, r2 uintptr, lastErr error)

Parameters are uintptr's, which are fine for integer values, but mainly we pass the addresses of the Go variables so the callee can read from and write into them.

Writing into mean dereferencing the pointers passed.
Thus, a *byte in Go is a char** in C.