No pointer size assumptions in the ABI boundary
Go will call us with uintptr's, which are unsigned and large enough to hold any pointer.
The equivalent for that is C99 uintptr_t (also C++11 and forward).
On pointers we are safe, but accepting uintptrs into a int32_t is a narrowing conversion.

To be more generic and future proof we typedef from intptr_t, preserving the integer value signal on non-pointer values.