Start the agent with a merged view of the env vars
The Environment class was a construct to ease testing the GUI components.
Sharing it with the agent at startup is useful for integration testing.
The GUI can talk to a sandboxed agent.
Usefulness is limited,
the WSL instances couldn't detect such env changes.

Still useful for integration testing
as long as the GUI doesn't need any WSL instance talking to the agent.