Forces the DLL to be built before the Go agent
Just adding a ProjectReference into the agent.vcxproj was not enough to
enforce the build order we needed.
Thus very often when building the appx from a clean env the DLL would
not be added to the package.

The proposed approach enforces first building whatever project is linked
to the agent.vcxproj via ProjectReference declaration and only build the
agent after that Target completes.

Another target was added (GetDepsOutputPaths) to compute the paths of
the assemblies that building the ProjectReference projects will generate
beforehand, so we can later refer to them instead of searching files
with globs as previously done.