Move task management to its own sub-package
This component was complex enough that having it in the same package as
worker started to become dangerous (because the worker could look at all
of the internals of the taskManager).

I made it its own subpackage to have a better defined public/private
interface. I made this subpackage internal to "worker" so that it stays
invisible to the rest of the project.