Fix agent api flutter test
This test watches the LocalAppData\UbuntuPro dir expecting the addr file
to show up. The new .csum file was interfering, so I filtered by the
changes speciffically to files containing the word addr.

Note that I cannot watch for full/path/to/addr because there is a race
here. The agent essentially does:

```
echo "$address" > "addr.new"
mv "addr.new" "addr"
```

The flutter Event watcher usually catches the first one only (and using
take 2, or filtering with endsWith(addr) fails). We could watch for the
.new file but I think that'd be depending on an implementation detail.

Anyways, the result of this commit is that this race persists, but
subscription.csum does not interfere.