The code we want to write at the high level
The purchase method must now be callback based
That means the result will be provided to the callback
That implies the method channel `result` must be a shared_ptr
Since we may need to access it in the exception handler as well.

Since fetching a product will be blocking
we need to call it on a background thread
Purchasing needs to start on the main (UI) thread, though

All exceptions are captured.

Note: The DefaultContext header must select what Context impl tol use.
More headers being included to be pedantic (don't accept transitive)
Rule: don't #include in cpp what my component header provides directly.