Sync stdio if attaching to a parent console app
By default, the Flutter main.cpp:

1. creates a new console if parent is a debugger.
2. or attaches to the parent console.

It does (2) assuming the parent is the flutter tool
(flutter run or flutter test - for example).
If its not, no console output would be visible.
A DeviceDesktopLoggerReader is be required to collect the std outputs.

This change adds detection for the case when the parent is a console shell
Thus, not the flutter tool.
We want to see the console output in that case.
It could well be the end-to-end test. ;)