Implements the Flutter automation for e2e
Assertions should be minimal.
Only one test scenario must be active for the lifetime of the process.
The CLI arg[0] determines what scenario to activate.
That matches the test function name on the Go side.
There is where the assertions must exist.
There we assert the side effects on registry, distros etc.
This is merely a script (compiled to machine code, ofc)