Makes agent recovery automatic if unreachable
When the GUI finds the addr file, but the agent doesn't respond
The GUI assumes the agent crashed or was killed some time in the past.

A button was offered to allow the user to request a recovery action.

The recovery action consists in:
- deleting the addr file and
- reset internal state to try again.

What changed: now that action is taken automatically.