Config: get rid of default registry values
All default values are empty strings, and the linter was also
complaining about readValue always receiving the same value for its
'defaultValue' argument.

Hence, I simplified the code so that there are no default values, or
rather the default values are always "". We can bring them back if we
ever need one of the defaults to be non-empty.