The main I want to write
GUI applications rely on some sort of event loop to mediate the
unpredictable timing user input will happen.

The Windows OS models this by means of the window message loop
and the WndProc function that says how the window responds to
messages it receives.

We want to wait on ReadFile to drain the console pipes, so the
child process won't block on the next printf after the pipe becomes
full.
We also want to wait on the agent process so we can quit if it quits.
We also want to respond to messages sent by the OS (such as WM_QUIT).

Threads without actual windows still have the message loop.
We can combine the window approach with some waiting functions and
create a custom message loop that allows our thread to sleep while
we don't receive any relevant input neither on the thread queue,
nor on the console pipe or the agent process termination.

Creating the pseudo-console should be declarative.
Starting the agent process under the pseudo-console should be a function
call.
Reacting to the agent's process termination or new input from the
console pipe should resemble reactive programming by passing functions.

I don't want to see tons of Win32 in my main()

I hope you enjoy this small spec.

:)