Making HRESULT into a usable exception type
Leveraging FormatMessage function to generate a string message
out of the system's catalog, based on the HRESULT value.
Plus some sugar with std::source_location.

That will come handy in avoiding freeing memory and closing handles
in multiple code paths.

A word on `unique_string`: Because FormatMessage allocates a buffer
and expects the caller to later free it using Windows Heap functions
this small wrapper based on std::unique_ptr makes it very handy to
deal with such confusing allocate/deallocate scenarios.

Logging: we don't do it everywhere, this is a very tiny application.
But if an exception occur, than we should.
Thus, LogSingleShot(): opens a file, writes to it and closes it.
In a single shot. Not the best performance, but meant to be called just
once, before quitting.