Implements the custom event loop
Although we never create or show a window, this is still a GUI
application
Thus it's powered by the window/thread message loop I mentioned before.
The need for the event loop might still not be clear enough.
Let me detail it:

1. Pseudo-console stdio HANDLEs are anonymous pipes HANDLEs.
2. The child process inherits the parts it needs to implement stdio.
3. The parent (this program) holds the other end of the pipes.
4. If stdout pipe becomes full, the child process will block at the next
   attempt to printf.
5. Thus we need to periodically read from the pipe to drain it.
6. We can do whatever we want with the bytes we read, nothing inclusive.

7. But we must read it.

8. Calling ReadFile on an empty pipe blocks the caller until new
   data arrives.
9. The child process might quit without writing anything new.

10. The parent will wait forever.

11. Users could kill this program (via task manager, for example)
12. That causes the process to receive WM_QUIT (amongst other related
    messages)

13. We need to respond to that message (and others maybe).

A lot of asynchrony.

MsgWaitForMultipleObjectsEx to the rescue. Seems made for this problem.
We can wait on multiple HANDLEs that the OS can interpret as
synchronizeable, i.e. that has some state that can change meaningfully,
such as a process that terminates or a pipe that receives new data.
Not to be confused with Overlapped IO. That's a whole other world.
It also waits on new messages arriving to the window or thread message
queue.
If configured accordingly it blocks until one of the interesting
activities happen, letting us know which one, so we can handle it
accordingly.

The association between which HANDLE to watch and which behavior to
adopt could be hardcoded in the event loop itself, but I found so
beautiful it passed as a parameter at the event loop construction.
Otherwise the event loop would be part of the PseudoConsole class,
making it like a God.