Write address file at the home directory
This directory is not virtualized. Moving the addr file here means
we will not need a virtualization exeption. The rest of the LocalAppData
contents need not be read by other applications, so they can stay there.

We also no longer need to create the cache user directory, we can assume
the home directory exists. Hence, so we can no longer test for it. I
removed that test case and added a new one so that we have at least
one error case.