Update E2E to use the new file locations
The E2E test cleans up the LocalAppData folder to avoid test coupling.
This commit updates this cleanup to the changes to the new path for the
address file.

We also watch that file to know if the agent is serving already. This
commit updates the path being watched.

The log finder now searches in the virtualized directory.