Follow usual convention for Apps
I changed the structure of the code so it more closely resembles the
WSL-Pro-Service and Windows-Agent apps.

The Landscape mock App matches the interface except that it is missing
the Quit method. I did not implement it because I saw no reason to. This
method is used only in signal handling, which we don't use in the
Landscape mock.