Improve error check in asyncRecv
We used to check that the message was not nill, but forgot to check the
error.

This double-check (is err nil? + is message nil?) would cause a bit of
ugly nesting so I moved it into newHostInfo so we can return instead of
nest.

This required a renaming of newHostInfo to receiveHostInfo to
keep the naming accurate.