Replaces placeholder text by real copy
Oliver suggested that text during the Engineering Sprint.
Ideally that should be clickable.
But let's not invest too much now because Elio is overhauling those screens.

As a consequence I'll have to rewrite most of the Flutter UI code
(the *_page.dart and *_widgets.dart files)

The parameterized l10n string, though, will be used in the final screen.