Give up on using fromJSON
It cannot be used by the action because action outputs must be strings.
It cannot be used by the consumer because in the following snippet:
```
${{ fromJSON(steps.load.outputs.data).flutter-version }}
```
the variable `steps.load.outputs.data` is not dereferenced, so it tries
to parse it as a JSON (and fails). So let's avoid being too smart and
use a plain file for this.