WizardRoute.onLoad to get updated SubscriptionInfo
SubscriptionStatusPage is sensitive to the current SubscriptionInfo
coming from the background agent.
WizardRoute.onLoad is a natural place to get this kind of information.
Everytime the wizard loads this route, a gRPC call is done
to fetch the latest information.
As a side effect the buildStartup local function becomes simpler.