Landscape: unify ini file parsing
We're about to add yet another location where we need to parse the file,
in order to obtain the account name and the registration key. This would
mean a third place where we parse the conf file. I decided to simply
have one function that readds all the config, even if it's a bit
redundant at times.