Landcape: Remove redundant test
Both tests "Success in first contact" and "Success with non-empty
config" are the same: the happy path. I removed the former because of
its misleading name: they all have a non-empty config. I also removed
its fixture.

I renamed the latter because the first contact thing does not mean much:
it's just the default happy path so "Success" is as good name as any.