Update tests to pass a mutex to distros
In production code, distros are only ever created by the database, so
the DB owns and distributes the mutex. In tests, however, many distros
are created.

If we create a global mutex for all tests, they can become slower
needlesly (as we run with mocks). If we create individual mutexes for
every subtest, we risk freezing up when running the tests without mocks
(this is something we were already exposed to).

The solution is to only use a global mutex when the distros are real,
and use individual mutexes when testing with mocks.