Fix debian package paths in E2E tests
For some reason the path being found by locateWSLProService. I fixed it
by getting rid of this function.

For more context: we were doing a weird mix of keeping track of the path
of the debian pkg in a local variable, and the path to the MSIX in a
global one. Instead, I moved the DEB to the global namespace for
consistency's sake. The code is now more readable.