Refactor Landscape Service
It is a bit ugly for proservices to have references to each other.
Instead, I split the proservices in two components:
- the service itself which is owned by the proservices.Manager. This
  component is in charge of managing the lifetime of the service.
- a service controler that can be passed around. This is a more
  restricted interface that allows only certain interactions.
  Proservices can own each other's controler, and that is how they
  interact with one another.

This may appear overkill for now, but soon enough the UI service will
also need to interact with the Landscape one (there'll be a GUI for
Landscape settings).

It is also planned that there will be a registry watcher service. This
service will need to interact with both Landscape amb Pro.

Hence, this split between the service itself and its controler. Note
that for now we only need to split the Landscape service, and we'll go
on splitting services as needed.