Simplify serviceConn interface
Method connection was a bit over-complicated, needing to lock a mutex
and return an unlock, as well as dangerous (as we could easily trigger a
deadlock by not notiing that this function locks).

Instead, I removed this function and added the only use-cases for it:
checking the connection status and sending info to the Landscape server.