Create taskQueue data structure
This data structure is a Queue that allows pushing and pulling tasks
from a FIFO queue, with the particularity that duplicated elements
will be removed in favour of the latest one.

Pulling from an empty queue will wait until it is no longer empty.

The queue is thread-safe, but there is no guarantee that awaiting
pulls will be served in order of arrival. This is OK because we
never pull more than one element at a time.