Avoid races in executing tasks in tests
Data modified during execution races with the marshalling in Save. This
is only an issue because these tasks are pointers, so all copies are shallow.
For non-pointer tasks, the copy being marshalled are deep so there is no
race to be had.

They need to be pointers because we want some special behaviour in
tests. Production tasks (Landscape and Pro Token) are not pointers but
rather regular structs, so they don't have to deal with this complexity.