Convert pseudo console stdout to Named pipe
Anonymous pipe doesn't support OVERLAPPED (async) IO
We need to async read from the stdout read HANDLE
Because the launcher process is getting currently blocked
on an infinite call to ReadFile.

We create the write handle on the same call,
thus when the constructor exits, we are ready to submit an async read
request.

Interestingly, PIPE_NOWAIT should not be used with named pipes and
Overlapped IO.