Implements the means of async reading hOutRead
Overlapped IO defers the act of reading the handle to a background
thread we don't see
When the reading is complete, the event member of the OVERLAPPED
structure is signaled, so the main thread can wait on it
as before, by calling MsgWaitForMultipleObjectsEx.
So the read operation is split in two parts:
- StartRead makes the request
- BytesRead returns the result
we don't really see the operation.

That enables the desired behavior: our main thread must be blocked most
of the time in the MsgWaitForMultipleObjectsEx call, nothing else.
This way it becomes trully reactive:
- it responds to new data arriving in the console pipe;
- it responds to the child process termination;
- and it responds to any message arriving in the Window queue.