Update existing tests and move some to controlstream_test.go
New can no longer fail as the connection to the agent is now done
during Serve.

This means that all its failure modes now move to Serve.

Serve retries intead of immediately serving, so we need some more
finesse in making assertions.

Many tests asserted on the state of the port file. This is now
responsability of the controlstream, so these have been moved.