Make system.UserProfileDir fail when the path does not exist
The daemon no longer exists upon failure, rather it retries. However it
makes no sense to retry if the entire Windows home directory is missing.

This will happen mostly when Windows filesystem mounting is disabled.
Re-enabling it requires a reboot, so it makes no sense for the daemon
to retry.

Hence, failing early seems like a good solution.