Fix bad check in daemon_test.go
This check succeeds:
    require.NotEqual(t, nil, net.ParseIP(address), "Address should be valid")

This check fails:
    require.NotNil(t, net.ParseIP(address), "Address should be valid")

Deep in the require->assert->reflect dependency chain there is a type comparison. Besides, the adress of type [::]:port
is not recognizable by net.ParseIP.

What we really want to assert is that the WSL-Pro-Service will be able
to use this address, so we make sure we can extract a port from it.