Fix erroneous test
The previous commit's refactor revealed a bug.

The config must reject introducing a token when a higher-priority one is
active. We were not loading the existing token, so we had an outdated
view of the current subscription. So when setting the user token, we
thought we had no active Store token, hence allowing it. There was even
a test asserting this missbehaviour!

This is fixed as of the previous commit, and this commit fixes the bad
test.