fix(windows-agent): Fix deadlock in config Stop
This commit fixes a deadlock between config.Stop and any config
notifaction callback that uses a config Mutex.

Scenario:

1. A config notification is triggered: the waitgroup count is increased
   by one and a single callback is called and starts executing.
   Waitgroup count decrease is deferred until the callback returns.

2. config.Stop is called, the config's ctx is cancelled but it is too late:
   the callback has already started. Then the config mutex is locked and
   it waits on the config waitgroup. This blocks until the waitgroup
   count reached 0.

3. during its execution, the callback calls some getter in the config.
   The mutex is already locked, so the getter cannot proceed, hence
   it cannot return. The deferred waitgroup count decrease never happens.

This is a deadlock! config.Stop is waiting on all callbacks to exit
before releasing the mutex, but the callback cannot exit until the mutex
is released.

Since waitgroups are thread-safe, we can simply remove the mutex from
config.Stop.