Fix Landscape's retrial time staircase
Every time Landscape fails to connect, the next attempt should wait for
twice as long. When a successful connection is made, this time duration
is reset to the minimum.

The reconnection cooldown was broken, and Landscpae was constantly
spamming the server (and the logs). This is now fixed.

Note: disable whitespace in this diff, as most lines only have a change
of indentation.