Watch parent key if UbuntuPro key does not exist
This will only happen if the user removes (or renames the UbuntuPro
key). In that case, we simply watch the HKCU/Software/ key, which is
almost guaranteed to exist.

I did not test this because it requires substantial mocking complexity
just to cover two lines of code.

I did test it manually.