Fix regression: not checking subscription status on start
For users it looks like the app forgotten the token previously supplied.
The current state is always starting in the subscribe now page.
When refactoring the app code, we mistakenly placed the call to get the current subscription status in the last screen,
it should be on the first.
Added an integration test to assert that behavior.