Better way to navigate from the subscription status page
This arrangement requires less knowledge about the existing routes
It makes sense to inquiry whether there is a previous page before requesting to go back.

Wsing Wizard.replace() on the startup page allows that trick, because that page will not count on the number of loaded pages.

Also, another advantage of using replace instead of next or jump is that we prevent the possibility of navigating back: the page is unloaded by the wizard router.