Remove the build-deb.ps1 script
This script made it easier to build the deb package on Windows, but now we have to copy the whole repository to give it a version name.
This makes it muuuuuch slower (takes like 5 minutes to rsync, same with just cp -r) so no one is going to use it. Let's avoid the maintenance burden by removing the file