Allows the AgentApiClient to change its connection
This way we can keep a stable client object even though
the underlying stub and channels change.
This prevents breaking the various view-models that depend on a stable reference to this API client.

We store a stub factory function so classes derived from UIClient can have theirt their preferred
creation behavior injected during tests.
In production it should always be UIClient.new.
This way we no longer need the AgentApiClient.withStub constructor.
When connecting to a new endpoint the stub is recreated using the same factory function.
A new test case is aded to certify that:
- connectTo doesn't change the object reference, just its internals;
- the internal stub is still a mock (indirectly asserted by the fact that ping succeeds even on port 9).