Restructured the tutorial to better capture the story: - Set things up (includes install up4w, configure up4w, approve registration for up4w host) - Watch UP4W transform your Ubuntu Pro game on WSL (with the remaining content structured into just 3 subsections focused on the main stories) - Tear things down (content to be added) - Next steps (added table guiding people on how to read the rest of the docs)
In the setup section, touched things up to better capture the basic steps and the caveats for where the user may want to skip.
In the same spirit, moved the wslconfig thing to a warning in the Install WSL section.

Formatted all button names in boldface. (Cf. our styleguide.)

Touched up a few other things for better flow.

Added deployment diagram.