Implements a status bar
Containing the version information on the left and
buttons to report a bug in GH and restart the agent if down.
We have strong assumptions about the URL to be launched to report a bug
Thus I made it testable by allowing replacement for url_launcher's launchUrl function.

Also, notice the usage of the Consumer widget (from provider package)
It makes provider and the provided AgentConnection components
depedencies of this one.