Create helper to run commands and capture output
The issue: we want only Stdout. However, in case of error, we also want Stderr in the error message.

We could use cmd.Output followed by an errors.As(err, &exitError) and then extract
the Stderr from the exit error.

But this is very tedious and we want to do it for every command, so let's sprinkle some sugar and make our lives easier with this helper.