Restrict landscape config file from 0604 to 0640
The landscape group needs read-access to the /etc/landscape/client.conf file.
However, we do not want any user to be able to read it as it may contain
confidential details such as the registration key. The solution is to restrict
its visibility to its owner and group, and to add it to landscape's group

Thanks @iosifache for spotting this vulnerability.