Avoid leaking proservices Manager on teardown
When we call agent.Quit, we don't wait for the proServices to
perform their cleanup. This means that the process can exit while the
Landscape, registry watcher or database are in the middle of an
operation.

This PR fixes this, somewhat: It guarantees to exit the process after
proservices.Stop returns. We trust that the implementation of the
teardown of these proservices is implemented properly.

This problem also affected tests, as the proServices could be writing to
disk while t.Cleanup tries to delete the temp directory, causing a race
and then a test error.