Fixes post CMake 3.29 and MSBuild 17.9
Something changed that gtest_discover_tests is now
breaking the unit test build process.
The option DISCOVERY_MODE PRE_TEST defers the problematic step to the run phase instead of the build phase.
Adding more verbosity and specifying the build config
to ctest as a precaution for the future
if something else breaks us