Agent: creates a auth token for gRPC messages
The token is written to the listener file as fixed length string
But it's passed to the gRPC services further encoded
One must know the encoding to be able to use the contents of the listener port file
Not a big secret, though, since the format can be guessed and this app is open source.
Yet it makes harder to attackers to issue malicious requests.