Agent API: New structure for the .address file
in the agent API proto so both wsl-pro-service and the GUI can reuse it
since both need to read that file
Not intended for sending via any RPC, just for serialization purposes.
We could have used other facilities, like YAML serialization in Go
But making it inside the proto defers the maintanance burden to
code generation.