Agent: implements creation of self-signed TLS certificates
Those certificates will grant gRPC a secure channel
In my tests locally Defender didn't complain about the certificates.
It's advisable to test in an environment similar to corporate to make sure it works as expected.

Clients only need the root CA certificate to trust the server certificate, with the caveat that the server name must be aliased to the CN declared in the certificate.