Ensures cloudinit.New returns an empty object on failure
What that does even mean?
That only returns an error if writeAgentData does so.
that could be for a number of errors, all serialization-related.
If we returned a usable object under such circunstance
and registered it to be notified of changes in the agent's config, silent failures would be logged on update, while the agent would stay ill-running.