Place the logic to find the WSL adapter IP in networking.go
Platform agnostic
Using the Option/defaultOptions idiom to determine:
 - the WSL CLI to invoke to determine WSL networking mode and
 - the platform behavior to create the linked list of network adapter
   addresses

The Windows implementation relies on the low level syscall.
That's the reason for unsafe buffer manipulation.

The reasoning for that choice remains the same:
- We cannot rely on the net package of stdlib because it doesn't return
  the adapter description
- Friendly Name can be changed by the user, so relying solely on that
  can be misleading.

For WSL networking mode we rely on executing `wsl --system`, mocked by
passing different command line invocation and environment.
Instead of implementing a type for that, we pass those as part of the
options struct.