This is a combination of 2 commits.
Fills the blanks with Windows specific behavior

A typedef to replace the windows.ipAdapterAddresses struct as that's not
available on Linux.
That allows the Linux version to implement a suitable replacement.
Very shallow wrapping around the low level data structure to expose
a nicer API returning strings and net.IP.
fillFromTemplate is a helper for the mocks to be able to construct the
right type during tests.