Add library annotations to please the up-to-date linter
Metadata such as `@TestOn` applied top-level as currently done could be ambiguous whether it refers to the current file (or library) or the import directive below it or whatever comes first after the metadata. Some kinds of metadata could be applicable to both, and it could matter (not the case of the `@TestOn` metadata, though).

https://dart.dev/tools/linter-rules/library_annotations