s.setBlocking race with the Apply* methods
As the handlers that invoke the Apply* methods run in other goroutines
even though the service.setBlocking is called before issuings RPCs
that will eventually be handled by the mockService, it may happen that
the goroutines could be running in different threads and thus have
no visibility of the changes to the s.blockingCalls flag.
Either using atomics or a RWMutex should suffice and make the test
future proof, as we may decide to change the order of the calls for
other reasons.